Support Python 3.12 and drop Python 3.7 in Core

Link to Issue: https://gitlab.com/tumult-labs/tumult/-/issues/3242

### For the Assignee:
- [X] Update the UNRELEASED sections of the changelogs for each project modified, 
        or explain in the MR description why no updates were necessary.
- [x] Check for any security implications of this MR and address them ([explanation](https://gitlab.com/tumult-labs/tumult/-/wikis/Developer-Tips-and-Tweaks/MRSecurityChecklist)).

### For the Reviewer:
- [ ] Confirm that an Issue is linked, if applicable.
- [ ] Confirm that the assignee updated relevant changelogs or provided appropriate
    justification as to why no updates were necessary.
- [ ] Check for any security implications of this MR and address them ([explanation](https://gitlab.com/tumult-labs/tumult/-/wikis/Developer-Tips-and-Tweaks/MRSecurityChecklist)).

This template lives [here](https://gitlab.com/tumult-labs/core/-/blob/dev/.gitlab/merge_request_templates/default_template.md)

Approved-by: Tom Magerlein <tom.magerlein@tmlt.io>
Approved-by: Amritha Pai <amritha.pai@tmlt.io>

See merge request tumult-labs/core!314

+++ test/unit/transformations/spark_transformations/test_agg.py
@@ -340,7 +340,7 @@ class TestCountGrouped(PySparkTest):
                 group_keys=group_keys,
             )
         ).toPandas()
-        expected_counts_df = pd.DataFrame({"C": [0]})
+        expected_counts_df = pd.DataFrame({"C": [0]}, dtype="int32")
         self.assert_frame_equal_with_sort(actual_counts_df, expected_counts_df)
 
     def test_empty_keys_but_nonempty_data(self):
